refactor(chain): consolidate finalized endpoint snapshot boundary
