+++ packages/chain/test/strict-current-finalized-evm-rpc.unit.test.ts
@@ -4,8 +4,6 @@ import {
 } from '@origintrail-official/dkg-core';
 import { afterEach, describe, expect, it } from 'vitest';
 
-import { snapshotStrictCurrentFinalizedEvmConfigV1 } from '../src/strict-current-finalized-evm-config.js';
-
 import {
   CONTROL_EIP1271_ATTEMPT_TIMEOUT_MS_V1,
   CONTROL_EIP1271_CALL_FROM_V1,
@@ -21,21 +19,15 @@ import {
 } from '../src/control-object-signature-verifier.js';
 import {
   CURRENT_FINALIZED_EVM_READ_ATTEMPT_TIMEOUT_MS_V1,
-  CURRENT_FINALIZED_EVM_READ_CALL_FROM_V1,
-  CURRENT_FINALIZED_EVM_READ_ENDPOINT_ATTEMPT_POLICY_V1,
-  CURRENT_FINALIZED_EVM_READ_GAS_LIMIT_V1,
-  CURRENT_FINALIZED_EVM_READ_MAX_ATTEMPTS_V1,
   CURRENT_FINALIZED_EVM_READ_MAX_CALLS_V1,
   CURRENT_FINALIZED_EVM_READ_MAX_CONCURRENT_PER_CHAIN_V1,
   CURRENT_FINALIZED_EVM_READ_MAX_RETURN_BYTES_V1,
   CURRENT_FINALIZED_EVM_READ_MAX_RPC_RESPONSE_BYTES_V1,
-  CURRENT_FINALIZED_EVM_READ_TOTAL_DEADLINE_MS_V1,
 } from '../src/current-finalized-evm-read-profile.js';
 import {
   createStrictCurrentFinalizedEvmChainAdapterV1,
   createStrictCurrentFinalizedEvmReadV1,
   readStrictCurrentFinalizedEvmRevertDataV1,
-  type StrictCurrentFinalizedEvmRpcConfigV1,
 } from '../src/strict-current-finalized-evm-rpc.js';
 import {
   executeStrictFinalizedAnchorPolicyV1,
@@ -69,23 +61,6 @@ afterEach(async () => {
 });
 
 describe('RFC-64 strict current-finalized raw JSON-RPC transport', () => {
-  it('keeps the EIP-1271 specialization pinned to the generic finalized-read profile', () => {
-    expect(CONTROL_EIP1271_CALL_FROM_V1).toBe(CURRENT_FINALIZED_EVM_READ_CALL_FROM_V1);
-    expect(CONTROL_EIP1271_GAS_LIMIT_V1).toBe(CURRENT_FINALIZED_EVM_READ_GAS_LIMIT_V1);
-    expect(CONTROL_EIP1271_MAX_RPC_RESPONSE_BYTES_V1)
-      .toBe(CURRENT_FINALIZED_EVM_READ_MAX_RPC_RESPONSE_BYTES_V1);
-    expect(CONTROL_EIP1271_ATTEMPT_TIMEOUT_MS_V1)
-      .toBe(CURRENT_FINALIZED_EVM_READ_ATTEMPT_TIMEOUT_MS_V1);
-    expect(CONTROL_EIP1271_MAX_ATTEMPTS_V1)
-      .toBe(CURRENT_FINALIZED_EVM_READ_MAX_ATTEMPTS_V1);
-    expect(CONTROL_EIP1271_MAX_CONCURRENT_CALLS_PER_CHAIN_V1)
-      .toBe(CURRENT_FINALIZED_EVM_READ_MAX_CONCURRENT_PER_CHAIN_V1);
-    expect(CONTROL_EIP1271_TOTAL_DEADLINE_MS_V1)
-      .toBe(CURRENT_FINALIZED_EVM_READ_TOTAL_DEADLINE_MS_V1);
-    expect(CONTROL_EIP1271_ENDPOINT_ATTEMPT_POLICY_V1)
-      .toBe(CURRENT_FINALIZED_EVM_READ_ENDPOINT_ATTEMPT_POLICY_V1);
-  });
-
   it('preserves an undefined generic result through an authenticated numbered anchor', async () => {
     const anchor = parseStrictFinalizedAnchorV1(
       { number: '0x7b', hash: BLOCK_HASH },
@@ -921,53 +896,7 @@ describe('RFC-64 strict current-finalized raw JSON-RPC transport', () => {
     expect(second.calls).toHaveLength(4);
   }, 12_000);
 
-  it('rejects unsafe configuration', () => {
-    const third = 'http://127.0.0.1:3';
-    for (const config of [
-      { chainId: '020430', endpoints: ['http://127.0.0.1:1'] },
-      { chainId: CHAIN_ID, endpoints: [] },
-      { chainId: CHAIN_ID, endpoints: ['ftp://127.0.0.1/a'] },
-      { chainId: CHAIN_ID, endpoints: ['http://127.0.0.1/a#fragment'] },
-      { chainId: CHAIN_ID, endpoints: ['https://user:pass@rpc.example.com'] },
-      { chainId: CHAIN_ID, endpoints: ['http://127.0.0.1:1'], peerEndpoint: third },
-    ]) {
-      expect(() => createStrictCurrentFinalizedEvmChainAdapterV1(
-        config as StrictCurrentFinalizedEvmRpcConfigV1,
-      )).toThrow(TypeError);
-    }
-  });
 
-  it('SELECTS the first two origins from a larger pool instead of rejecting it', () => {
-    // Behaviour change, and the point of the change. This case previously sat in
-    // the rejection list above — which is precisely why RFC64's finalized-VM
-    // precommit could not construct a scope on any shipped EVM network, where
-    // `resolveRpcUrls(rpcUrl, rpcUrls)` yields three URLs.
-    //
-    // The two-endpoint ATTEMPT ceiling is unchanged; selection just makes sure
-    // no more than two ever reach it. The dedup-then-ceiling test above still
-    // pins that ceiling.
-    // Assert WHICH two, not merely that it constructed — "returns a function"
-    // would pass just as green if selection kept the last two, or one, or all.
-    // Driven through the CONFIG boundary rather than the selector directly: the
-    // config is what derives provider identity, so this exercises the real
-    // pipeline instead of hand-built records.
-    const selection = snapshotStrictCurrentFinalizedEvmConfigV1({
-      chainId: CHAIN_ID,
-      endpoints: ['http://127.0.0.1:1', 'http://127.0.0.1:2', 'http://127.0.0.1:3'],
-    } as never).endpoints;
-    // Distinct ports are distinct origins, so this session really does carry two
-    // providers — and the third is dropped, not merged.
-    expect(selection).toEqual(['http://127.0.0.1:1/', 'http://127.0.0.1:2/']);
-
-    expect(() => createStrictCurrentFinalizedEvmChainAdapterV1({
-      chainId: CHAIN_ID,
-      endpoints: [
-        'http://127.0.0.1:1',
-        'http://127.0.0.1:2',
-        'http://127.0.0.1:3',
-      ],
-    } as StrictCurrentFinalizedEvmRpcConfigV1)).not.toThrow();
-  });
 });
 
 interface SuccessfulHandlerOptions {
